# Conflicts:
#	cfg/gm_msg.xml
#	cfg/sample_script.xml
#	cli_common/dftCliMsgHandler.cpp
#	cli_common/dftCliMsgHandler.h
#	include/_error_def.h
#	plugin_u3d/apoClientU3d.cpp

+++ apoBluePrint/include/apoScript/listdialog.h
@@ -23,7 +23,7 @@ public:
     }
     bool GetSelectText(QString &val) {
 
-        if(-1==m_selIndex || m_selIndex >= m_selList.size()) {
+        if(-1==m_selIndex || m_selIndex >= (int)m_selList.size()) {
             return false ;
         }
         val = m_selList[m_selIndex] ;
